requirements.txt
streamlit==1.28.1
pandas==2.1.1
plotly==5.17.0
openpyxl==3.1.2
requests==2.31.0
numpy==1.25.2
yfinance==0.2.28
gspread==5.12.0
google-auth==2.23.4 